в main.js оставить только вызов функции getComments,
